feat: 상품 API 코드 일부 모듈화
