+++ express-mission-prisma/src/services/productService.js
@@ -1,4 +1,8 @@
 import productRepositpry from "../repositories/productRepositpry.js";
+import {
+  createKeywordfilterOtions,
+  createPageSizefilterOptions,
+} from "../utils/product/productFilterOptions.js";
 
 async function create(createData) {
   return await productRepositpry.create(createData);
@@ -9,55 +13,13 @@ async function getById(id) {
 }
 
 async function getAllByfilter(query) {
-  const { page, pageSize, orderBy, keyWord = "" } = query;
-
-  const pageNum = page || 1;
-  const pageSizeNum = pageSize || 10;
-  const order = orderBy || "recent";
-  const offset = (pageNum - 1) * pageSizeNum;
-  const whereOrBody = {
-    contains: keyWord,
-    mode: "insensitive",
-  };
-  const whereOr = {
-    OR: [
-      {
-        name: whereOrBody,
-      },
-      {
-        description: whereOrBody,
-      },
-    ],
-  };
-
-  const filterOptions = {
-    orderBy: { createdAt: "desc" },
-    skip: parseInt(offset),
-    take: parseInt(pageSizeNum),
-    where: whereOr,
-  };
-
-  return await productRepositpry.getAllByfilter(filterOptions);
+  const filterByPageSize = createPageSizefilterOptions(query);
+  return await productRepositpry.getAllByfilter(filterByPageSize);
 }
 
 async function countByfilter(query) {
-  const { keyWord = "" } = query;
-  const whereOrBody = {
-    contains: keyWord,
-    mode: "insensitive",
-  };
-  const whereOr = {
-    OR: [
-      {
-        name: whereOrBody,
-      },
-      {
-        description: whereOrBody,
-      },
-    ],
-  };
-
-  return await productRepositpry.countByfilter(whereOr);
+  const filterByKeword = createKeywordfilterOtions(query);
+  return await productRepositpry.countByfilter(filterByKeword);
 }
 
 async function update(id, updateData) {
