+++ express-mission-prisma/src/utils/product/prepareProductData.js
