+++ express-mission-prisma/src/utils/product/productFilterOptions.js
@@ -1,0 +1,54 @@
+function createPageSizefilterOptions(query) {
+  const { page, pageSize, orderBy, keyWord = "" } = query;
+
+  const pageNum = page || 1;
+  const pageSizeNum = pageSize || 10;
+  const order = orderBy || "recent";
+  const offset = (pageNum - 1) * pageSizeNum;
+  const whereOrBody = {
+    contains: keyWord,
+    mode: "insensitive",
+  };
+  const whereOr = {
+    OR: [
+      {
+        name: whereOrBody,
+      },
+      {
+        description: whereOrBody,
+      },
+    ],
+  };
+
+  const filterOptions = {
+    orderBy: { createdAt: "desc" },
+    skip: parseInt(offset),
+    take: parseInt(pageSizeNum),
+    where: whereOr,
+  };
+
+  return filterOptions;
+}
+
+function createKeywordfilterOtions(query) {
+  const { keyWord = "" } = query;
+
+  const filterBody = {
+    contains: keyWord,
+    mode: "insensitive",
+  };
+  const filterOptions = {
+    OR: [
+      {
+        name: filterBody,
+      },
+      {
+        description: filterBody,
+      },
+    ],
+  };
+
+  return filterOptions;
+}
+
+export { createPageSizefilterOptions, createKeywordfilterOtions };
